refactor toolbar buttonGroup & direct return linkchooserpreviewflyout
